backend/apps/news/serializers.py
from rest_framework import serializers # type: ignore
from .models import Article

class ArticleSerializer(serializers.ModelSerializer):
    class Meta:
        model = Article
        fields = '__all__'
